Ejecución del cálculo

+++ APPWEB-INCOMEMotorCalculoJava-api/src/main/java/com/inditex/rrhh/icmclcwb/api/app/dto/JobDto.java
@@ -1,11 +1,13 @@
 package com.inditex.rrhh.icmclcwb.api.app.dto;
 
-import javax.validation.constraints.NotNull;
-import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
 import java.util.Date;
+
+import javax.validation.constraints.NotNull;
+
+import io.swagger.annotations.ApiModelProperty;
 
 @Data
 public class JobDto implements Serializable {
@@ -13,7 +15,7 @@ public class JobDto implements Serializable {
 	private static final long serialVersionUID = -5607507260079003307L;
 
 	@ApiModelProperty(value = "Identificador del trabajo", required = false)
-	private Integer id;
+	private Long id;
 
 	@ApiModelProperty(value = "Id del páis a ejecutar", required = false)
 	private String idPais;
